Fix: #2135, segmentation-faults when working with file channels (#2141)
* Add missing method to SeekableByteChannel interface. Create test case.
* Provide basic implementation for Channels.new{Input, Output}Stream
* Remove stub annotations from Channels, remain method declarations to be defined in the future
