+++ javalib/src/main/scala/java/nio/channels/SeekableByteChannel.scala
@@ -6,6 +6,7 @@ trait SeekableByteChannel extends ByteChannel {
   override def read(dst: ByteBuffer): Int
   override def write(src: ByteBuffer): Int
   def position(): Long
+  def position(newPosition: Long): SeekableByteChannel
   def size(): Long
   def truncate(size: Long): SeekableByteChannel
 }
